\name{fcFilter-methods}
\docType{methods}
\alias{fcFilter-methods}
\title{ Feature-Condition Filter}
\description{
  Feature-Condition Filter for biclusters
}
\section{Methods}{
\describe{

\item{\code{signature(object = "Biclust")}}{
  Use \code{help("fcFilter")} see help and examples
}
}}
\keyword{methods}

